+++ demos/manual/src/main/java/is/codion/framework/demos/manual/common/demo/ApplicationPanel.java
@@ -82,7 +82,7 @@ public final class ApplicationPanel extends JPanel {
 
     JPanel inputPanel = flexibleGridLayoutPanel(0, 2).build();
 
-    textField(model.shortStringValue())
+    Components.stringField(model.shortStringValue())
             .columns(20)
             .lowerCase(true)
             .maximumLength(20)
@@ -303,7 +303,7 @@ public final class ApplicationPanel extends JPanel {
     add(inputPanel, BorderLayout.CENTER);
 
     flexibleGridLayoutPanel(2, 1)
-            .add(textField()
+            .add(stringField()
                     .columns(20)
                     .editable(false)
                     .focusable(false)