+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/Components.java
@@ -444,7 +444,7 @@ public final class Components {
    * @param <B> the builder type
    * @return a JTextField builder
    */
-  public static <B extends TextFieldBuilder<String, JTextField, B>> TextFieldBuilder<String, JTextField, B> textField() {
+  public static <B extends TextFieldBuilder<String, JTextField, B>> TextFieldBuilder<String, JTextField, B> stringField() {
     return TextFieldBuilder.builder(String.class);
   }
 
@@ -453,7 +453,7 @@ public final class Components {
    * @param <B> the builder type
    * @return a JTextField builder
    */
-  public static <B extends TextFieldBuilder<String, JTextField, B>> TextFieldBuilder<String, JTextField, B> textField(Value<String> linkedValue) {
+  public static <B extends TextFieldBuilder<String, JTextField, B>> TextFieldBuilder<String, JTextField, B> stringField(Value<String> linkedValue) {
     return TextFieldBuilder.builder(String.class, linkedValue);
   }
 
@@ -930,7 +930,7 @@ public final class Components {
    * @return a {@link FileInputPanel} builder
    */
   public static FileInputPanel.Builder fileInputPanel() {
-    return fileInputPanel(textField()
+    return fileInputPanel(stringField()
             .editable(false)
             .focusable(false)
             .build());