+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/ColumnConditionPanel.java
@@ -348,7 +348,7 @@ public final class ColumnConditionPanel<C, T> extends JPanel {
                 .build();
       }
       else if (columnClass.equals(String.class)) {
-        return textField((Value<String>) linkedValue)
+        return stringField((Value<String>) linkedValue)
                 .build();
       }