+++ swing/framework-server-monitor/src/main/java/is/codion/swing/framework/server/monitor/ui/ServerMonitorPanel.java
@@ -285,7 +285,7 @@ public final class ServerMonitorPanel extends JPanel {
   }
 
   private JTextField createMemoryField() {
-    return textField()
+    return stringField()
             .columns(8)
             .editable(false)
             .horizontalAlignment(SwingConstants.CENTER)